@@ -1,7 +1,7 @@
-package leetcode.tests;
+package leetcode;
 
 import leetcode.Sqrt;
-import org.junit.Test;
+import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
